:truck: chore(api): move `certifiable-profile-for-learning-content-repository` to `src`
